Updated with Jagadish's suggestions.
