+++ samza-core/src/test/java/org/apache/samza/table/caching/TestCachingTable.java
@@ -35,10 +35,10 @@ import org.apache.samza.table.ReadWriteTable;
 import org.apache.samza.table.ReadableTable;
 import org.apache.samza.table.TableSpec;
 import org.apache.samza.table.caching.guava.GuavaCacheTable;
-import org.apache.samza.table.descriptors.caching.CachingTableDescriptor;
-import org.apache.samza.table.descriptors.caching.CachingTableProvider;
-import org.apache.samza.table.descriptors.caching.guava.GuavaCacheTableDescriptor;
-import org.apache.samza.table.descriptors.caching.guava.GuavaCacheTableProvider;
+import org.apache.samza.table.caching.descriptors.CachingTableDescriptor;
+import org.apache.samza.table.caching.descriptors.CachingTableProvider;
+import org.apache.samza.table.caching.guava.descriptors.GuavaCacheTableDescriptor;
+import org.apache.samza.table.caching.guava.descriptors.GuavaCacheTableProvider;
 import org.apache.samza.table.remote.RemoteReadWriteTable;
 import org.apache.samza.table.remote.TableRateLimiter;
 import org.apache.samza.table.remote.TableReadFunction;
