+++ samza-test/src/test/java/org/apache/samza/test/table/TestTableDescriptorsProvider.java
@@ -41,8 +41,8 @@ import org.apache.samza.storage.kv.descriptors.RocksDbTableDescriptor;
 import org.apache.samza.storage.kv.descriptors.RocksDbTableProviderFactory;
 import org.apache.samza.table.TableConfigGenerator;
 import org.apache.samza.table.TableDescriptorsProvider;
-import org.apache.samza.table.descriptors.remote.RemoteTableDescriptor;
-import org.apache.samza.table.descriptors.remote.RemoteTableProviderFactory;
+import org.apache.samza.table.remote.descriptors.RemoteTableDescriptor;
+import org.apache.samza.table.remote.descriptors.RemoteTableProviderFactory;
 import org.apache.samza.table.remote.TableReadFunction;
 import org.apache.samza.util.RateLimiter;
 import org.apache.samza.util.Util;
